Using a SpellQuery object in our searching services.
